stockcli: command line utility to query stock data using ticker symbol
StockCLI is a command line utility to get historic prices of a Stock using it's
ticker symbol. It uses Yahoo Finance to source data but the plan is extend to
include other sources. The date range can be specified in multiples of days,
weeks, months or range with a start and end date.

Example usages to get Google stock prices:
1. stockli.py GOOG "today"
2. stockli.py GOOG "5d"
3. stockli.py GOOG "2w"
4. stockli.py GOOG "1m"
5. stockli.py GOOG "01-10-2017 to 10-10-2017"

Signed-off-by: Arun Siluvery <arunk.siluvery@gmail.com>